Focus release status and publish helpers
Use the exact ASCII-only status vocabulary, reduce website notes refresh to a single dispatch, and move publish source and duplicate checks into the pre-confirmation workflow. Replace the queue helper with a compact build-number-only script and focused tests.


Copilot-Session: ae9429ab-3288-4306-bef2-334a88751a86
